requirements.txt
anthropic
fastapi
flask
flask-cors
gunicorn
numpy
openai
Pillow
pillow-heif
python-dotenv
python-multipart
scikit-learn
spacy
uvicorn
